requirements.txt
streamlit
pandas
plotly
Pillow
matplotlib
seaborn
numpy
scikit-learn
base64